refactor(workspace): extract command palette catalog search
